backend/requirements/base.txt
beautifulsoup4==4.12.3
boto3==1.28.52
botocore==1.31.52
celery==5.3.1
click==8.1.7
cookiecutter==2.3.0
crispy-bootstrap5==0.7
django==4.2.11
django-axes==6.4.0
django-cachalot==2.6.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-decorator-include==3.0
django-environ==0.11.2
django-formtools==2.4.1
django-ipware==7.0.1
django-phonenumber-field==7.1.0
django-redis==5.3.0
django-rest-knox==4.2.0
django-session-security==2.6.7
django-smtp-ssl==1.0
django-storages==1.14
django-tenants==3.6.1
djangorestframework==3.15.2
flake8==6.1.0
loguru==0.7.2
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-boto3sqs==0.43b0
opentelemetry-instrumentation-botocore==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==24.1
phonenumberslite==8.13.23
pluginbase==1.0.1
pre-commit==3.8.0
psycopg2-binary==2.9.9
pydub==0.25.1
PyJWT==2.8.0
python3-saml==1.16.0
pytz==2024.1
setuptools==70.1.1
XlsxWriter==3.1.9
